Refactor cachix config into a postBuildStep
It's not necessary to keep the cachix configuration as a special python module key. As the config gets translated into a post_build_step in python, we can just as easily do that in nix, and reduce the layers of indirection that somebody (say, somebody adding a new caching system) has to look through.